Seperated app.py into different modules for seperation of tasks and organisation
