+++ app/config.py
@@ -1,0 +1,13 @@
+import os
+
+class Config:
+    # Secret key for session logic
+    secret_key = os.getenv('SECRET_KEY', 'dev-secret-key')  # swap for real env var later
+    # Configure the database URI and initialize the database
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///app.db'
+
+
+    # Moved from app.py but 
+    # "it is in general a good practice to set configuration from environment variables, 
+    # and provide a fallback value when the environment does not define the variable" - From Flask Mega Tutorial (my saviour)
+    # Maybe something to think about when handling databases (Thanks flask mega tutorial)
